+++ CHANGELOG.md
@@ -29,6 +29,9 @@
 - Retrieve and show all data for found Items [#1465](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1465)
 - Added toggle to retrieve all data for found Items [#1466](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1466)
 
+### `ImpEx` enhancements
+- Introduced `In-Editor Results` view [#1467](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1467)
+
 ### `Groovy` enhancements
 - Introduced `batch` execution mode for Groovy files [#1450](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1450)